feat: Error page node type
Error page configuration is complex and cumbersome. Especially when code
and configuration is deployed, adding a new error page means creating
the page node in the live environment, adding settings accordingly to
the codebase and deploying the configuration afterward.

This feature allows editors to create page nodes on any level on the
page tree, select the error codes this error page should respond to.
The newly created page node now responds to every error occurring on
sibling pages of the error pages or below.